+++ src/hooks/api/useFileDownload.js
@@ -5,7 +5,7 @@ import { useErrorMessages } from '../useErrorMessages';
 export const QUERY_KEY_DOWNLOAD_LOGS = 'downloadLogs';
 export const QUERY_KEY_DOWNLOAD_ACTION_MENU = 'downloadActionMenu';
 export const QUERY_KEY_DOWNLOAD_PREVIEW_MODAL = 'downloadPreviewModal';
-export const QUERY_KEY_DOWNLOAD_MARK_PREVIEW_MODAL = 'downloadMarkPreviewModal';
+export const QUERY_KEY_DOWNLOAD_MARC_PREVIEW_MODAL = 'downloadMarcPreviewModal';
 
 export const useFileDownload = ({
   id,